pyproject.toml
[project]
name = "a2a-sk"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "asyncclick>=8.1.8",
    "asyncio>=3.4.3",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "jwcrypto>=1.5.6",
    "pyjwt>=2.10.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.2.1",
    "starlette>=0.46.1",
    "uvicorn>=0.34.0",
    "semantic-kernel>=1.28.0"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
